requirements.txt
pandas
mysql-connector-python
nba_api
dotenv
streamlit
seaborn
matplotlib
plotly
